Add asynchronous closure arguments example
This adds example code for using visiting_ref to work around current
limitations with how lifetimes can be expressed for functions that take
an async closure as an argument, passing a reference with an internal
lifetime to the closure and running the closure within the function.

To make this example cleaner, the individual "futures" crate
dev-dependencies have been replaced with just the main futures crate,
with example code in doc comments updated accordingly. A table of
contents has also been added to the top-level crate documentation to
make it easier to jump to various sections, as the documentation is
fairly lengthy.