Disable "needless_doctest_main" Clippy lint
Clippy issues a "needless_doctest_main" warning for our example code
that uses the "#[tokio::main]" attribute. Since we need to define a main
function to use the Tokio attribute, the lint warning should be ignored.
The warning has been disabled for all doc tests in the crate, as I'm
unsure how to disable it for just the relevant code block (placing the
attribute on the main function itself doesn't seem to work, nor does
placing it at the top of the code block in a crate attribute).